requirements.txt
Django==5.2
djangorestframework==3.14.0
psycopg2-binary==2.9.9
Pillow==10.2.0
django-cors-headers==4.3.1
python-dotenv==1.0.1
whitenoise==6.6.0
gunicorn==21.2.0

# For AI features
nltk>=3.8.1
spacy>=3.7.2
# Optional: better performance for spaCy
# en-core-web-sm>=3.7.0

# Transformers and dependencies are optional with the new AI implementation
# transformers==4.38.1 
# torch>=2.0.0
# sentencepiece>=0.1.99
# accelerate>=0.22.0

# Platform-specific packages (uncomment as needed)
# For Windows: already includes psycopg2-binary

# For macOS (uncommment if needed)
# psycopg2-binary can be used for macOS too, or if there are issues:
# psycopg2==2.9.9